@@ -1,46 +1,98 @@
 {
-  "name": "angular-electron",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+  "name": "dManager",
+  "version": "0.4.0",
+  "description": "A Manager, installer and updater for Minecraft 1.13 datapacks.",
+  "homepage": "http://stevertus.ga/dManager",
+  "author": {
+    "name": "Stevertus",
+    "email": "stevertusyt@gmail.com"
   },
+  "keywords": [
+    "angular",
+    "angular 5",
+    "electron",
+    "typescript",
+    "sass"
+  ],
+  "main": "main.js",
   "private": true,
+  "scripts": {
+    "postinstall": "npx electron-builder install-app-deps",
+    "ng": "ng",
+    "start": "node hooks/environments/set_profile.js && npm-run-all -p ng:serve electron:serve",
+    "build": "node hooks/environments/set_profile.js && ng build && npm run electron:tsc",
+    "build:prod": "node hooks/environments/set_profile.js && ng build --prod && npm run electron:tsc",
+    "ng:serve": "ng serve -o",
+    "electron:tsc": "tsc main.ts",
+    "electron:serve": "wait-on http-get://localhost:4200/ && npm run electron:tsc && electron . --serve",
+    "electron:local": "npm run build && electron .",
+    "electron:linux": "npm run build:prod && npx electron-builder build --linux",
+    "electron:windows": "npm run build:prod && npx electron-builder build --windows",
+    "electron:mac": "npm run build:prod && npx electron-builder build --mac",
+    "test": "karma start ./karma.conf.js",
+    "pree2e:build": "webdriver-manager update --gecko false && npm run ng serve",
+    "pree2e:protractor": "wait-on http-get://localhost:4200/ && protractor ./protractor.conf.js",
+    "e2e": "npm-run-all -p -r pree2e:build pree2e:protractor"
+  },
   "dependencies": {
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4"
+    "@covalent/code-editor": "^1.0.0",
+    "@nebular/theme": "^2.0.0-rc.9",
+    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
+    "blob-to-buffer": "^1.2.8",
+    "bootstrap": "^4.1.1",
+    "file-saver": "^1.3.3",
+    "jquery": "^3.3.1",
+    "jszip": "^3.1.5",
+    "jszip-utils": "0.0.2",
+    "nebular-icons": "^1.0.6",
+    "ngx-fs": "0.0.1",
+    "ngx-markdown": "^1.5.2",
+    "ngx-monaco-editor": "^4.1.1",
+    "popper.js": "^1.14.3"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~2.0.0",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
-    "karma-cli": "~1.0.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "protractor": "~5.1.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "@angular/cli": "1.7.4",
+    "@angular/common": "5.2.9",
+    "@angular/compiler": "5.2.9",
+    "@angular/compiler-cli": "5.2.9",
+    "@angular/core": "5.2.9",
+    "@angular/forms": "5.2.9",
+    "@angular/http": "5.2.9",
+    "@angular/language-service": "5.2.9",
+    "@angular/platform-browser": "5.2.9",
+    "@angular/platform-browser-dynamic": "5.2.9",
+    "@angular/router": "5.2.9",
+    "@ngx-translate/core": "9.1.1",
+    "@ngx-translate/http-loader": "2.0.1",
+    "@types/core-js": "0.9.36",
+    "@types/jasmine": "2.8.6",
+    "@types/jasminewd2": "2.0.3",
+    "@types/node": "7.0.7",
+    "codelyzer": "4.2.1",
+    "core-js": "2.5.5",
+    "cross-env": "5.1.4",
+    "dotenv": "5.0.1",
+    "electron": "1.8.4",
+    "electron-builder": "20.8.1",
+    "electron-reload": "1.2.2",
+    "fs-extra": "5.0.0",
+    "jasmine-core": "2.99.1",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "2.0.0",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-coverage-istanbul-reporter": "1.4.2",
+    "karma-jasmine": "1.1.1",
+    "karma-jasmine-html-reporter": "0.2.2",
+    "npm-run-all": "4.1.2",
+    "npx": "10.0.1",
+    "protractor": "5.3.1",
+    "replace": "0.3.0",
+    "rxjs": "5.5.8",
+    "ts-node": "4.1.0",
+    "tslint": "5.9.1",
+    "typescript": "2.6.2",
+    "wait-on": "2.1.0",
+    "webdriver-manager": "12.0.6",
+    "zone.js": "0.8.26"
   }
 }
