@@ -1,4 +1,5 @@
 import { AngularElectronPage } from './app.po';
+import { browser, element, by } from 'protractor';
 
 describe('angular-electron App', () => {
   let page: AngularElectronPage;
@@ -7,8 +8,8 @@ describe('angular-electron App', () => {
     page = new AngularElectronPage();
   });
 
-  it('should display message saying app works', () => {
-    page.navigateTo();
-    expect(page.getParagraphText()).toEqual('app works!');
+  it('should display message saying App works !', () => {
+    page.navigateTo('/');
+    expect(element(by.css('app-home h1')).getText()).toMatch('App works !');
   });
 });
