@@ -1,3 +1,4 @@
-export const environment = {
-  production: true
+export const CONF_PROD = {
+  production: true,
+  environment: 'PROD'
 };
