@@ -1,5 +1,11 @@
 /* SystemJS module definition */
-declare var module: NodeModule;
+declare var nodeModule: NodeModule;
 interface NodeModule {
   id: string;
+}
+
+declare var window: Window;
+interface Window {
+  process: any;
+  require: any;
 }
